# docs/swagger.yaml
basePath: /
definitions:
  handlers.LoginUserBody:
    properties:
      email:
        type: string
      password:
        maxLength: 50
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  handlers.RegisterUserBody:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      password:
        maxLength: 50
        minLength: 8
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - confirmPassword
    - email
    - password
    - username
    type: object
...
info:
  contact:
    email: nirazshaw156@gmail.com
    name: Niraj Shaw
    url: https://github.com/Niraj1910
  description: Secure task management API with Golang/Gin Framework, JWT authentication
    and email verification.
  title: Golang Task REST API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns JWT token in cookie
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginUserBody'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - Auth
...
  /register:
    post:
      consumes:
      - application/json
      description: Creates a pending registration record and sends a verification
        email.
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterUserBody'
      produces:
      - application/json
      responses:
        "201":
          description: Registration request received. Check email for verification
            link
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Email already registered
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /verify:
    get:
      consumes:
      - application/json
      description: Validates the token and email, creates the real user, and deletes
        the temp verification record.
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      - description: User email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Account verified and activated
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid or expired token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Verification record not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify email and complete registration
      tags:
      - Auth
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token (format: Bearer <token>)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"